requirements.txt
ipykernel
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-groq
python-dotenv
faiss-cpu
torch
sentence-transformers
arxiv
wikipedia
streamlit
duckduckgo-search